♻️ Refactor: 상수,타입 파일분리, dependancy initialSchedules 삭제
