+++ src/types/activities.ts
@@ -1,0 +1,5 @@
+export interface ScheduleBase {
+  date: string;
+  startTime: string;
+  endTime: string;
+}
